TPK9N0T extract App.tsx's issue mutations and swimlane modals into hooks
